pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linecoach"
version = "0.1.0"
description = "Multi-agent Coaching and Real-time Escalation system for customer service interactions"
readme = "README.md"
authors = [
    {name = "Capgemini"}
]
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-speech>=2.0.0",
    "pyaudio>=0.2.11",
    "numpy>=1.20.0",
    "vertexai>=1.0.0",
    "google-auth>=2.0.0",
    "google-cloud-core>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.3.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.24.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pre-commit>=4.1.0",
    "ruff>=0.9.2"
]

[project.urls]
"Homepage" = "https://github.com/noelabu/LineCoach"
"Bug Tracker" = "https://github.com/noelabu/LineCoach/issues"

[tool.setuptools]
packages = ["app"]
